Change number type to be `f64`, as per the lua standard.
https://www.lua.org/pil/2.3.html

We still represent indices as i32, if they fit into one. This is just an
optimisation. In the future we can do the same for values.

+++ src/ast.rs
@@ -39,7 +39,7 @@ pub enum Expression {
 
 #[derive(Debug, Clone)]
 pub enum Term {
-    Number(i32),
+    Number(f64),
     String(String),
     Boolean(bool),
     Variable(String),
